+++ src/Servidor/Modelo/magnetman.h
@@ -9,7 +9,7 @@ class Arma_fireman;
 class Magnetman: public Personaje_npc{
 	private:
 		Mapa *mapa;
-		size_t tiempo_pasado;
+		float tiempo_pasado;
 		Arma_fireman* arma;
     bool atacando;
 	public:
@@ -19,11 +19,11 @@ class Magnetman: public Personaje_npc{
 		//Dado una Mapa* y una direccion, ataca.
 		virtual void atacar(int dir, Mapa* mapa);
 		//Dado un tiempo pasado y un Mapa*, se mueve.
-		virtual void mover(size_t tiempo, Mapa* mapa);
+		virtual void mover(float tiempo, Mapa* mapa);
 		//Dada una Bala*, recibe un ataque.
 		virtual void recibir_ataque(Bala* ataque);
 		//Dado un tiempo, realiza un update.
-		virtual void update(size_t tiempo);
+		virtual void update(float tiempo);
 };
 
 #endif //MAGNETMAN_H
