+++ src/Servidor/Modelo/strategy_mover_jumping_snipper.cpp
@@ -3,7 +3,7 @@
 
 #define TIEMPO_SALTO 4
 
-void StrategyMoverJumpingSnipper::mover(Mapa *mapa, Jumping_snipper *pj, size_t tiempo){
+void StrategyMoverJumpingSnipper::mover(Mapa *mapa, Jumping_snipper *pj, float tiempo){
 	tiempo_pasado += tiempo;
 	Coordenada nueva_coord(-1, -1);
 	if (tiempo_pasado >= TIEMPO_SALTO){
